make QM molecule partial sanitized with RDKit
This aligns the RDKit conversion process. A relaxed sanitization
process is required to avoid kekulization/sanitization/valence issues
which would prevent a molecule from being created.

Especially relevant in the context of `draw`, which has an RDKit
backend that calls this function. We don't want it to fail drawing
simple because it doesn't follow the sanitization rules.