Change to_rdkit_mol bond handling. Add ignore_bond_order option.
Now that we use RDKit to detect rings, we create a lot of rdkit molecules
in which we don't care about the bond orders, sometimes we have non-integer
or weird bond orders, and we also don't want to waste time sanitizing them
since all we care about is connectivity of the molecular graph.
So if you know you're calling to_rdkit_mol in this setting, call it with
ignore_bond_orders=True and sanitize=False.