When doing ring detection, don't pass bond orders to RDKit.
This should save complication and time. Also, don't sanitize.
All we're going to ask it for is ring information. 
For that all bonds (that we give it) count.
If we want it to ignore H-bonds, or vdW bonds, for some scenarios,
then we should figure that out. But often we use this ring detection 
code for things like drawing, which does want to use those bonds.