Simplify and optimize cutting label lookup in to_rdkit_mol.
As far as I can tell, cutting labels can only be "L" or "R".
I had a previous commit that checked this is equivalent, and 
by running the unit tests and regression tests could not find
a counterexample.
This check is much simpler, and about 200 times faster, than
the previous call to detect_cutting_label.
Also simplified the dictionary structure.