Fragment.to_rdkit_mol now respects some kwargs instead of printing warnings.
Previously, if you call to_rdkit_mol on a Fragment, it would always return
the mapping, but if you called it on a Molecule, it would by default not.

Now if you call it with return_mapping=False (or not specified) then it 
doesn't return the mapping, and if you call it with return_mapping=True,
then it does.

Internally, it calls converter.to_rdkit_mol with return_mapping=True,
but doesn't pass it back unless asked. Now the behavior of thing.to_rdkit_mol()
is the same whether thing is a Molecule or a Fragment (which is a subclass
of Molecule). 
Should reduce number of warnings, eg. when drawing fragments.

Also cleaned up the code a bit, and added some docstrings,